requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.1.8
distro==1.9.0
exceptiongroup==1.3.0
fastapi==0.115.12
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
multidict==6.4.4
openai==1.86.0
propcache==0.3.2
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.4
sniffio==1.3.1
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.4.0
uvicorn==0.34.3
yarl==1.20.1
langchain>=0.1.14
langchain-openai>=0.1.3
chromadb>=0.4.24
